feat: middlewares and validation of companies routers

+++ src/services/companies/delete-company.service.ts
@@ -1,10 +1,21 @@
+import AppError from "../../errors/app-error";
 import { prisma } from "../../prisma/client"
 
 const deleteCompanyService = async (id: number) => {
 
-    await prisma.license.deleteMany({ where: { companyId: id } });
+    const company = await prisma.company.findUnique({ where: { id }});
+    if(!company){
+        throw new AppError(404, "Company not found");
+    }
 
-    return prisma.company.delete({ where: { id }});
+    const deleteLicense = prisma.license.deleteMany({ where: { companyId: id } });
+    const deleteCompany = prisma.company.delete({ where: { id }});
+
+    try {
+        await prisma.$transaction([ deleteLicense, deleteCompany ]);
+    } catch (error) {
+        throw new AppError(403, "Error while deleting company")
+    }
 }
 
 export default deleteCompanyService;
